All ints now are long (in datastore and bprimitivetype, redis 2

+++ Utilities/BCloudServiceUtilities-BMemoryService-Redis/BMemoryServiceRedis.cs
@@ -61,7 +61,7 @@ namespace BCloudServiceUtilities.MemoryServices
             {
                 if (AsDouble % 1 == 0)
                 {
-                    return new BPrimitiveType((int)AsDouble);
+                    return new BPrimitiveType((long)AsDouble);
                 }
                 return new BPrimitiveType(AsDouble);
             }
